ジェネリクスを使用することによるDB周りのミスを修正
