+++ src/main/java/jp/gmo/ojt/chachat/domain/model/CanvasDrawHistory.java
@@ -1,14 +1,6 @@
 package jp.gmo.ojt.chachat.domain.model;
 
-import java.io.Serializable;
-import java.util.Date;
-
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.PrePersist;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
@@ -17,15 +9,7 @@ import lombok.Data;
 @Data
 @Entity
 @SuppressWarnings("serial")
-public class CanvasDrawHistory implements Serializable{
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
-    @NotNull
-	@Embedded
-    private CanvasIdentity canvasIdentity;
-	@NotBlank
-	private String targetId;
+public class CanvasDrawHistory extends CanvasHistory{
 	@NotBlank
 	private String mode;
 	@NotBlank
@@ -41,38 +25,6 @@ public class CanvasDrawHistory implements Serializable{
 	@NotNull
 	private Integer toY;
 	
-    @NotNull
-    private Date createdAt;
-
-    
-    public CanvasIdentity getCanvasIdentity() {
-    	return this.canvasIdentity;
-    }
-    public void setCanvasIdentity(CanvasIdentity canvasIdentity) {
-    	this.canvasIdentity = canvasIdentity;
-    }
-    
-    
-	public String getRoomId() {
-		return this.canvasIdentity.getRoomId();
-	}
-	public void setRoomId(String roomId) {
-		this.canvasIdentity.setRoomId(roomId);
-	}
-	
-	public Integer getCanvasId() {
-		return this.canvasIdentity.getCanvasId();
-	}
-	public void setCanvasId(Integer canvasId) {
-		this.canvasIdentity.setCanvasId(canvasId);
-	}
-	
-	public String  getTargetId() {
-		return this.targetId;
-	}
-	public void setTargetId(String targetId) {
-		this.targetId = targetId;
-	}
 	public String  getMode() {
 		return this.mode;
 	}
@@ -114,9 +66,5 @@ public class CanvasDrawHistory implements Serializable{
 	public void setToY(Integer toY) {
 		this.toY = toY;
 	}
-    
-	@PrePersist
-	protected void onCreate() {
-	  this.createdAt = new Date();
-	}
+
 }
