+++ src/main/java/jp/gmo/ojt/chachat/domain/model/CanvasObjectCreatedHistory.java
@@ -1,14 +1,6 @@
 package jp.gmo.ojt.chachat.domain.model;
 
-import java.io.Serializable;
-import java.util.Date;
-
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.PrePersist;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
@@ -17,53 +9,16 @@ import lombok.Data;
 @Data
 @Entity
 @SuppressWarnings("serial")
-public class CanvasObjectCreatedHistory implements Serializable{
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
-    @NotNull
-	@Embedded
-    private CanvasIdentity canvasIdentity;
-	@NotBlank
-	private String targetId;
+public class CanvasObjectCreatedHistory extends CanvasHistory{
 	@NotBlank
 	private String objectType;
 	@NotNull
 	private Integer x;
 	@NotNull
 	private Integer y;
-    @NotNull
-    private Date createdAt;
 
-    
-    public CanvasIdentity getCanvasIdentity() {
-    	return this.canvasIdentity;
-    }
-    public void setCanvasIdentity(CanvasIdentity canvasIdentity) {
-    	this.canvasIdentity = canvasIdentity;
-    }
-    
-    
-	public String getRoomId() {
-		return this.canvasIdentity.getRoomId();
-	}
-	public void setRoomId(String roomId) {
-		this.canvasIdentity.setRoomId(roomId);
-	}
-	
-	public Integer getCanvasId() {
-		return this.canvasIdentity.getCanvasId();
-	}
-	public void setCanvasId(Integer canvasId) {
-		this.canvasIdentity.setCanvasId(canvasId);
-	}
-	
-	public String  getTargetId() {
-		return this.targetId;
-	}
-	public void setTargetId(String targetId) {
-		this.targetId = targetId;
-	}
+
+
 	public String  getObjectType() {
 		return this.objectType;
 	}
@@ -83,8 +38,4 @@ public class CanvasObjectCreatedHistory implements Serializable{
 		this.y = y;
 	}
 
-	@PrePersist
-	protected void onCreate() {
-	  this.createdAt = new Date();
-	}
 }
