+++ src/main/java/jp/gmo/ojt/chachat/domain/repository/CanvasHistoryRepository.java
@@ -6,11 +6,11 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import jp.gmo.ojt.chachat.domain.model.CanvasDrawHistory;
+import jp.gmo.ojt.chachat.domain.model.CanvasHistory;
 import jp.gmo.ojt.chachat.domain.model.CanvasIdentity;
 
 @Repository
-public interface CanvasHistoryRepository<E> extends JpaRepository<E, CanvasIdentity>{
+public interface CanvasHistoryRepository<E extends CanvasHistory> extends JpaRepository<E, CanvasIdentity>{
 	List<E> findByCanvasIdentity(CanvasIdentity canvasIdentity);
 	List<E> findByCanvasIdentityAndCreatedAtBefore(CanvasIdentity canvasIdentity,Date date);
 	List<E> findByCanvasIdentityAndCreatedAtAfter(CanvasIdentity canvasIdentity,Date date);
