+++ src/main/java/jp/gmo/ojt/chachat/domain/service/CanvasHistoryService.java
@@ -3,15 +3,18 @@ package jp.gmo.ojt.chachat.domain.service;
 import java.util.Date;
 import java.util.List;
 
+import javax.annotation.Resource;
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import jp.gmo.ojt.chachat.domain.model.CanvasDrawHistory;
+import jp.gmo.ojt.chachat.domain.model.CanvasHistory;
 import jp.gmo.ojt.chachat.domain.model.CanvasIdentity;
 import jp.gmo.ojt.chachat.domain.repository.CanvasHistoryRepository;
 
 @Service
-public class CanvasHistoryService<E> {
+public class CanvasHistoryService<E extends CanvasHistory> {
 	@Autowired
 	CanvasHistoryRepository<E> canvasHistoryRepository;
 	
@@ -26,7 +29,7 @@ public class CanvasHistoryService<E> {
 	public List<E> getBeforeHistoriesByDate(CanvasIdentity canvasIdentity, Date date){
 		return canvasHistoryRepository.findByCanvasIdentityAndCreatedAtBefore(canvasIdentity, date);
 	}
-	
+
 	public void saveHistory(E canvasHistory) {
 		canvasHistoryRepository.save(canvasHistory);
 	}
